Update falsepositive.list
Hello,
I am reaching out regarding the domain flugger.pl, which has been categorized as suspicious/malicious. We believe this classification is incorrect and was caused by the shared IP address 20.50.2.30 in the Azure cloud environment. This IP is shared with other unrelated websites, including 29techhelpnjmk.azurewebsites.net, which has a risk rating of 9.21.
Flugger.pl is a legitimate website with no malicious intent or activity. We kindly request a review and removal of our domain from the blacklist to prevent any disruptions for our users. Please let us know if any additional information is required to support this request.

Signed-off-by: pawmopm <131968861+pawmopm@users.noreply.github.com>